fixing infinite loop bug and deleted menuheight calculation
