+++ package.json
@@ -26,6 +26,7 @@
     "react-dom": "^18.2.0",
     "react-popper": "^2.3.0",
     "react-scripts": "5.0.1",
+    "rollup-plugin-import-css": "^3.2.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
     "typescript": "^4.9.5",
@@ -73,6 +74,7 @@
     "nodemon": "^2.0.22",
     "rollup": "^2.79.1",
     "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-typescript2": "^0.34.1"
   },
   "peerDependencies": {
