+++ src/index.ts
@@ -1,5 +1,2 @@
-export * from "./slashMenuDisplay/display";
-export * from "./slashMenuDisplay/defaults";
-export * from "./slashMenuDisplay/ListItem";
-export * from "./slashMenuDisplay/utils";
-// import { ListItem } from "./slashMenuDisplay/ListItem";
+export * from "./slashMenuReact/display";
+export * from "./slashMenuReact/defaults";
